+++ string.c
@@ -90,11 +90,8 @@ VALUE rb_cSymbol;
  *            The string is not embedded. When a string is embedded, the contents
  *            follow the header. When a string is not embedded, the contents is
  *            on a separately allocated buffer.
- * 2:     STR_CHILLED_LITERAL (will be frozen in a future version)
+ * 2:     STR_CHILLED (will be frozen in a future version)
  *            The string was allocated as a literal in a file without an explicit `frozen_string_literal` comment.
- *            It emits a deprecation warning when mutated for the first time.
- * 3:     STR_CHILLED_SYMBOL_TO_S (will be frozen in a future version)
- *            The string was allocated by the `Symbol#to_s` method.
  *            It emits a deprecation warning when mutated for the first time.
  * 4:     STR_PRECOMPUTED_HASH
  *            The string is embedded and has its precomputed hashcode stored
@@ -2035,7 +2032,7 @@ rb_ec_str_resurrect(struct rb_execution_context_struct *ec, VALUE str, bool chil
         str_duplicate_setup_heap(klass, str, new_str);
     }
     if (chilled) {
-        FL_SET_RAW(new_str, STR_CHILLED_LITERAL);
+        FL_SET_RAW(new_str, STR_CHILLED);
     }
     return new_str;
 }
@@ -2046,7 +2043,7 @@ rb_str_with_debug_created_info(VALUE str, VALUE path, int line)
     VALUE debug_info = rb_ary_new_from_args(2, path, INT2FIX(line));
     if (OBJ_FROZEN_RAW(str)) str = rb_str_dup(str);
     rb_ivar_set(str, id_debug_created_info, rb_ary_freeze(debug_info));
-    FL_SET_RAW(str, STR_CHILLED_LITERAL);
+    FL_SET_RAW(str, STR_CHILLED);
     return rb_str_freeze(str);
 }
 
@@ -12441,9 +12438,7 @@ sym_inspect(VALUE sym)
 VALUE
 rb_sym_to_s(VALUE sym)
 {
-    VALUE str = str_new_shared(rb_cString, rb_sym2str(sym));
-    FL_SET_RAW(str, STR_CHILLED_SYMBOL_TO_S);
-    return str;
+    return rb_sym2str(sym);
 }
 
 VALUE